minidlna: force fixup of autotools.
Since we know we update the makefile.am do a manual autoreconf
instead of allowing this to be detected during make.

If its detected the make tries to use the original version of automake
and not the version installed on the HOST.

this is so we can build on ubuntu 15.10 without downgrading automake.