Add support for git mirror.
This is so that we can build even when remote git server
is down.

It is possible to use other git mirror than git@public.inteno.se
but the sanity checks will fail and it won't be possible
to mirror the inteno git repos as the mirror is disabled
for inteno servers.